+++ packages/devtools-reps/src/object-inspector/tests/object-inspector.js
@@ -12,341 +12,6 @@ const { MODE } = require("../../reps/constants");
 const { Rep } = require("../../reps/rep");
 const gripStubs = require("../../reps/stubs/grip");
 const accessorStubs = require("../../reps/stubs/accessor");
-const performanceStubs = require("../stubs/performance");
-
-const {
-  createNode,
-  getChildren,
-  getPromiseProperties,
-  getValue,
-  isDefault,
-  isPromise,
-  makeNodesForProperties,
-} = require("../utils");
-
-const root = {
-  path: "root",
-  contents: { value: {} }
-};
-
-const objProperties = {
-  ownProperties: {
-    "0": {
-      value: {}
-    },
-    "2": {},
-    length: {
-      value: 3
-    }
-  },
-  prototype: {
-    type: "object",
-    actor: "server2.conn1.child1/pausedobj618",
-    class: "bla"
-  }
-};
-
-describe("makeNodesForProperties", () => {
-  it("kitchen sink", () => {
-    const nodes = makeNodesForProperties(objProperties, root);
-
-    const names = nodes.map(n => n.name);
-    expect(names).toEqual(["0", "length", "__proto__"]);
-
-    const paths = nodes.map(n => n.path);
-    expect(paths).toEqual(["root/0", "root/length", "root/__proto__"]);
-  });
-
-  it("includes getters and setters", () => {
-    const nodes = makeNodesForProperties(
-      {
-        ownProperties: {
-          foo: { value: "foo" },
-          bar: {
-            "configurable": true,
-            "enumerable": true,
-            "get": {
-              "type": "object",
-              "actor": "server2.conn1.child1/obj106",
-              "class": "Function",
-              "extensible": true,
-              "frozen": false,
-              "sealed": false,
-              "name": "get x",
-              "displayName": "get x",
-              "location": {
-                "url": "debugger eval code",
-                "line": 1
-              }
-            },
-            "set": {
-              "type": "undefined"
-            }
-          },
-          baz: {
-            "configurable": true,
-            "enumerable": true,
-            "get": {
-              "type": "undefined"
-            },
-            "set": {
-              "type": "object",
-              "actor": "server2.conn1.child1/obj116",
-              "class": "Function",
-              "extensible": true,
-              "frozen": false,
-              "sealed": false,
-              "name": "set x",
-              "displayName": "set x",
-              "location": {
-                "url": "debugger eval code",
-                "line": 1
-              }
-            }
-          }
-        },
-        prototype: {
-          class: "bla"
-        }
-      },
-      root
-    );
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["bar", "baz", "foo", "__proto__"]);
-    expect(paths).toEqual(["root/bar", "root/baz", "root/foo", "root/__proto__"]);
-  });
-
-  it("sorts keys", () => {
-    const nodes = makeNodesForProperties(
-      {
-        ownProperties: {
-          bar: { value: {} },
-          1: { value: {} },
-          11: { value: {} },
-          2: { value: {} },
-          _bar: { value: {} }
-        },
-        prototype: {
-          class: "bla"
-        }
-      },
-      root
-    );
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["1", "2", "11", "_bar", "bar", "__proto__"]);
-    expect(paths).toEqual([
-      "root/1",
-      "root/2",
-      "root/11",
-      "root/_bar",
-      "root/bar",
-      "root/__proto__"
-    ]);
-  });
-
-  it("prototype is included", () => {
-    const nodes = makeNodesForProperties(
-      {
-        ownProperties: {
-          bar: { value: {} }
-        },
-        prototype: { value: {}, class: "bla" }
-      },
-      root
-    );
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["bar", "__proto__"]);
-    expect(paths).toEqual(["root/bar", "root/__proto__"]);
-  });
-
-  it("window object", () => {
-    const nodes = makeNodesForProperties(
-      {
-        ownProperties: {
-          bar: { value: {} },
-          location: { value: {} }
-        },
-        class: "Window"
-      },
-      {
-        path: "root",
-        contents: { value: { class: "Window" } }
-      }
-    );
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["bar", "[default properties]"]);
-    expect(paths).toEqual(["root/bar", "root/##-default"]);
-  });
-
-  it("window prop on normal object", () => {
-    const windowRoots = [
-      {
-        contents: { value: { class: "Window" } }
-      }
-    ];
-
-    const objectRoots = [
-      {
-        contents: { value: { class: "Object" } }
-      }
-    ];
-
-    const item = { name: "location" };
-
-    expect(isDefault(item, windowRoots)).toEqual(true);
-    expect(isDefault(item, objectRoots)).toEqual(false);
-  });
-
-  // For large arrays
-  it("numerical buckets", () => {
-    let objProps = { ownProperties: {}, prototype: { class: "Array" } };
-    for (let i = 0; i < 331; i++) {
-      objProps.ownProperties[i] = { value: {} };
-    }
-    const nodes = makeNodesForProperties(objProps, root);
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual([
-      "[0..99]",
-      "[100..199]",
-      "[200..299]",
-      "[300..331]",
-      "__proto__"
-    ]);
-
-    expect(paths).toEqual([
-      "root/bucket1",
-      "root/bucket2",
-      "root/bucket3",
-      "root/bucket4",
-      "root/__proto__"
-    ]);
-  });
-
-  it("quotes property names", () => {
-    const nodes = makeNodesForProperties(
-      {
-        ownProperties: {
-          // Numbers are ok.
-          332217: { value: {} },
-          "needs-quotes": { value: {} },
-          unquoted: { value: {} },
-          "": { value: {} }
-        },
-        prototype: {
-          class: "WindowPrototype"
-        }
-      },
-      root
-    );
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual([
-      '""',
-      "332217",
-      '"needs-quotes"',
-      "unquoted",
-      "__proto__"
-    ]);
-    expect(paths).toEqual([
-      "root/",
-      "root/332217",
-      "root/needs-quotes",
-      "root/unquoted",
-      "root/__proto__"
-    ]);
-  });
-});
-
-describe("promises utils function", () => {
-  it("is promise", () => {
-    const promise = {
-      contents: {
-        enumerable: true,
-        configurable: false,
-        value: {
-          frozen: false,
-          ownPropertyLength: 0,
-          preview: {
-            kind: "Object",
-            ownProperties: {},
-            ownPropertiesLength: 0,
-            safeGetterValues: {}
-          },
-          actor: "server2.conn2.child1/pausedobj36",
-          promiseState: {
-            state: "rejected",
-            reason: {
-              type: "undefined"
-            },
-            creationTimestamp: 1486584316133.3994,
-            timeToSettle: 0.001713000237941742
-          },
-          class: "Promise",
-          type: "object",
-          extensible: true,
-          sealed: false
-        },
-        writable: true
-      }
-    };
-
-    expect(isPromise(promise)).toEqual(true);
-  });
-
-  it("getPromiseProperties", () => {
-    const promise = {
-      path: "root",
-      contents: {
-        enumerable: true,
-        configurable: false,
-        value: {
-          frozen: false,
-          ownPropertyLength: 0,
-          preview: {
-            kind: "Object",
-            ownProperties: {},
-            ownPropertiesLength: 0,
-            safeGetterValues: {}
-          },
-          actor: "server2.conn2.child1/pausedobj36",
-          promiseState: {
-            state: "rejected",
-            reason: {
-              type: "3"
-            },
-            creationTimestamp: 1486584316133.3994,
-            timeToSettle: 0.001713000237941742
-          },
-          class: "Promise",
-          type: "object",
-          extensible: true,
-          sealed: false
-        },
-        writable: true
-      }
-    };
-
-    const properties = getPromiseProperties(promise);
-    expect(properties).toMatchSnapshot();
-  });
-});
 
 describe("ObjectInspector", () => {
   it("renders as expected", () => {
@@ -704,201 +369,5 @@ describe("ObjectInspector", () => {
     label.simulate("click");
 
     expect(onLabelClick.mock.calls.length).toBe(1);
-  });
-});
-
-describe("getChildren", () => {
-  it("accessors - getter", () => {
-    const nodes = getChildren({
-      item: createNode("root", "rootpath", accessorStubs.get("getter"))
-    });
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["<get>"]);
-    expect(paths).toEqual(["rootpath/get"]);
-  });
-
-  it("accessors - setter", () => {
-    const nodes = getChildren({
-      item: createNode("root", "rootpath", accessorStubs.get("setter"))
-    });
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["<set>"]);
-    expect(paths).toEqual(["rootpath/set"]);
-  });
-
-  it("accessors - getter & setter", () => {
-    const nodes = getChildren({
-      item: createNode("root", "rootpath", accessorStubs.get("getter setter"))
-    });
-
-    const names = nodes.map(n => n.name);
-    const paths = nodes.map(n => n.path);
-
-    expect(names).toEqual(["<get>", "<set>"]);
-    expect(paths).toEqual(["rootpath/get", "rootpath/set"]);
-  });
-
-  it("uses the expected actor to get properties", () => {
-    const stub = performanceStubs.get("performance");
-    const getObjectProperties = jest.fn();
-
-    // Test that the function gets the actor from the value.
-    getChildren({
-      actors: {},
-      item: createNode("root", "rootpath", stub.ownProperties.timing),
-      getObjectProperties
-    });
-    expect(getObjectProperties.mock.calls[0][0]).toBe("server2.conn4.child1/obj44");
-
-    // Test that the function gets the actor from the getterValue.
-    getChildren({
-      actors: {},
-      item: createNode("root", "rootpath", stub.safeGetterValues.timing),
-      getObjectProperties
-    });
-    expect(getObjectProperties.mock.calls[1][0]).toBe("server2.conn4.child1/obj44");
-  });
-
-  it("safeGetterValues", () => {
-    const stub = performanceStubs.get("timing");
-    const nodes = getChildren({
-      actors: {},
-      item: createNode("root", "rootpath", {
-        value: {
-          actor: "rootactor",
-          type: "object"
-        }
-      }),
-      getObjectProperties: actor => {
-        if (actor === "rootactor") {
-          return stub;
-        }
-        return null;
-      }
-    });
-
-    const nodeEntries = nodes.map(n => [n.name, getValue(n)]);
-    const nodePaths = nodes.map(n => n.path);
-
-    const childrenEntries = [
-      ["connectEnd", 1500967716401],
-      ["connectStart", 1500967716401],
-      ["domComplete", 1500967716719],
-      ["domContentLoadedEventEnd", 1500967716715],
-      ["domContentLoadedEventStart", 1500967716696],
-      ["domInteractive", 1500967716552],
-      ["domLoading", 1500967716426],
-      ["domainLookupEnd", 1500967716401],
-      ["domainLookupStart", 1500967716401],
-      ["fetchStart", 1500967716401],
-      ["loadEventEnd", 1500967716720],
-      ["loadEventStart", 1500967716719],
-      ["navigationStart", 1500967716401],
-      ["redirectEnd", 0],
-      ["redirectStart", 0],
-      ["requestStart", 1500967716401],
-      ["responseEnd", 1500967716401],
-      ["responseStart", 1500967716401],
-      ["secureConnectionStart", 1500967716401],
-      ["unloadEventEnd", 0],
-      ["unloadEventStart", 0],
-      ["__proto__", stub.prototype]
-    ];
-    const childrenPaths = childrenEntries.map(([name]) => `rootpath/${name}`);
-
-    expect(nodeEntries).toEqual(childrenEntries);
-    expect(nodePaths).toEqual(childrenPaths);
-  });
-});
-
-describe("getValue", () => {
-  it("get the value from contents.value", () => {
-    let item = {
-      contents: {
-        value: "my value"
-      }
-    };
-    expect(getValue(item)).toBe("my value");
-
-    item = {
-      contents: {
-        value: 0
-      }
-    };
-    expect(getValue(item)).toBe(0);
-
-    item = {
-      contents: {
-        value: false
-      }
-    };
-    expect(getValue(item)).toBe(false);
-
-    item = {
-      contents: {
-        value: null
-      }
-    };
-    expect(getValue(item)).toBe(null);
-  });
-
-  it("get the value from contents.getterValue", () => {
-    let item = {
-      contents: {
-        getterValue: "my getter value"
-      }
-    };
-    expect(getValue(item)).toBe("my getter value");
-
-    item = {
-      contents: {
-        getterValue: 0
-      }
-    };
-    expect(getValue(item)).toBe(0);
-
-    item = {
-      contents: {
-        getterValue: false
-      }
-    };
-    expect(getValue(item)).toBe(false);
-
-    item = {
-      contents: {
-        getterValue: null
-      }
-    };
-    expect(getValue(item)).toBe(null);
-  });
-
-  it("get the value from getter and setter", () => {
-    let item = {
-      contents: {
-        get: "get"
-      }
-    };
-    expect(getValue(item)).toEqual({get: "get"});
-
-    item = {
-      contents: {
-        set: "set"
-      }
-    };
-    expect(getValue(item)).toEqual({set: "set"});
-
-    item = {
-      contents: {
-        get: "get",
-        set: "set"
-      }
-    };
-    expect(getValue(item)).toEqual({get: "get", set: "set"});
   });
 });
